move auth middleware out of index and add hash/salt checking
