+++ functions/index.js
@@ -1,18 +1,15 @@
 const express = require('express');
 const cors = require('cors');
-const basicAuth = require('basic-auth');
 
 // The Cloud Functions for Firebase SDK to create Cloud Functions and set up triggers.
 const functions = require('firebase-functions');
 
 const {
-  firestore,
   fail,
-  isUserType,
-  STATUS_CODE_UNAUTHORIZED,
   STATUS_CODE_INTERNAL_SERVER_ERROR,
   ABLY_API_KEY_CUSTOMERS,
   ABLY_API_KEY_RIDERS,
+  HASHING_SECRET,
 } = require('./common');
 
 const {
@@ -24,59 +21,15 @@ const {
   getAbly,
 } = require('./Handlers');
 
-const { logger } = functions;
+const { authorizeMiddleware } = require('./auth');
 
 const SECRET_NAMES = [
   ABLY_API_KEY_RIDERS,
   ABLY_API_KEY_CUSTOMERS,
+  HASHING_SECRET,
   'MAPBOX_ACCESS_TOKEN',
   'GOOGLE_MAPS_API_KEY',
 ];
-
-const authorizeMiddleware = async (req, res, next) => {
-  const credentials = basicAuth(req);
-  if (!credentials) {
-    logger.info('Invalid Authorization header, or header missing.');
-    fail(res, STATUS_CODE_UNAUTHORIZED); // intentionally not providing a message as it could assist probing hackers
-    return;
-  }
-
-  const { name, pass } = credentials;
-
-  let data;
-  try {
-    const snapshot = await firestore.collection('users').doc(name).get();
-    data = snapshot.data();
-  } catch (error) {
-    // Ask Express to respond with HTTP 500 Internal Server Error.
-    // see: https://expressjs.com/en/guide/error-handling.html
-    // It's worth noting that, by default, Express serves this as HTML and includes all Error detail.
-    // It's also worth noting that this code was written against Express 4, given Express 5 was still in Beta - which
-    // explains why we need to do this error propagation 'by hand'.
-    next(error);
-    return;
-  }
-
-  if (!data) {
-    logger.info(`User '${name}' not found.`);
-  } else if (data.password === pass) {
-    const userType = data.type;
-    if (!isUserType(userType)) {
-      next(new Error(`User type '${userType}', as specified for user '${name}', is unknown.`));
-      return;
-    }
-
-    // make available for other middleware, specifically API handlers
-    res.locals.username = name;
-    res.locals.userType = userType;
-
-    next(); // Success
-    return;
-  }
-
-  logger.info(`Incorrect password supplied for user '${name}'. Received '${pass}', expected '${data.password}'.`);
-  fail(res, STATUS_CODE_UNAUTHORIZED); // intentionally not providing a message as it could assist probing hackers
-};
 
 const errorHandlingMiddleware = (err, req, res, next) => {
   if (res.headersSent) {
